Cast arguments to trig functions to doubles to prevent substitution of float functions.
When presented with code like:
atan2(float a, float b)
the compiler substitutes the float versions of the trig functions - eg:
atan2f(float a, float b)
These functions do not exist on th cRIO. Casing arguments to doubles prevents this.

+++ mpu6050/MPU6050_6Axis_MotionApps20.h
@@ -33,6 +33,7 @@ THE SOFTWARE.
 #ifndef _MPU6050_6AXIS_MOTIONAPPS20_H_
 #define _MPU6050_6AXIS_MOTIONAPPS20_H_
 
+#include "math.h"
 #include "I2Cdev.h"
 #include "helper_3dmath.h"
 
@@ -639,18 +640,23 @@ uint8_t MPU6050::dmpGetGravity(VectorFloat *v, Quaternion *q) {
 // uint8_t MPU6050::dmpGetEIS(long *data, const uint8_t* packet);
 
 uint8_t MPU6050::dmpGetEuler(double *data, Quaternion *q) {
-    data[0] = atan2(2*q -> x*q -> y - 2*q -> w*q -> z, 2*q -> w*q -> w + 2*q -> x*q -> x - 1); // psi
-    data[1] = -asin(2*q -> x*q -> z + 2*q -> w*q -> y); // theta
-    data[2] = atan2(2*q -> y*q -> z - 2*q -> w*q -> x, 2*q -> w*q -> w + 2*q -> z*q -> z - 1); // phi
+    data[0] = atan2(static_cast<double>(2*q -> x*q -> y - 2*q -> w*q -> z),
+        static_cast<double>(2*q -> w*q -> w + 2*q -> x*q -> x - 1)); // psi
+    data[1] = -asin(static_cast<double>(2*q -> x*q -> z + 2*q -> w*q -> y)); // theta
+    data[2] = atan2(static_cast<double>(2*q -> y*q -> z - 2*q -> w*q -> x),
+        static_cast<double>(2*q -> w*q -> w + 2*q -> z*q -> z - 1)); // phi
     return 0;
 }
 uint8_t MPU6050::dmpGetYawPitchRoll(double *data, Quaternion *q, VectorFloat *gravity) {
     // yaw: (about Z axis)
-    data[0] = atan2(2*q -> x*q -> y - 2*q -> w*q -> z, 2*q -> w*q -> w + 2*q -> x*q -> x - 1);
+    data[0] = atan2(static_cast<double>(2*q -> x*q -> y - 2*q -> w*q -> z),
+        static_cast<double>(2*q -> w*q -> w + 2*q -> x*q -> x - 1));
     // pitch: (nose up/down, about Y axis)
-    data[1] = atan(gravity -> x / sqrt(gravity -> y*gravity -> y + gravity -> z*gravity -> z));
+    data[1] = atan(static_cast<double>(gravity -> x 
+        / sqrt(gravity -> y*gravity -> y + gravity -> z*gravity -> z)));
     // roll: (tilt left/right, about X axis)
-    data[2] = atan(gravity -> y / sqrt(gravity -> x*gravity -> x + gravity -> z*gravity -> z));
+    data[2] = atan(static_cast<double>(gravity -> y 
+        / sqrt(gravity -> x*gravity -> x + gravity -> z*gravity -> z)));
     return 0;
 }
